+++ scripts/istio/deploy-istio.sh
@@ -27,12 +27,14 @@ kubectl apply -f "$DIR/istio-ns.yaml"
 kubectl apply -f ../private-resources/aspenmesh-istio-private-pr-pull-secret.yaml \
   --namespace istio-system
 
-"$DIR/generate-ca-cert.sh"
-kubectl create secret generic cacerts -n istio-system --from-file=ecc/ca-cert.pem \
-  --from-file=ecc/ca-key.pem --from-file=ecc/root-cert.pem \
-  --from-file=ecc/cert-chain.pem \
-  --dry-run -o yaml |
-  kubectl apply -f -
+if [[ ${UPDATE_CA_CERT:-} != "false" ]]; then
+  "$DIR/generate-ca-cert.sh"
+  kubectl create secret generic cacerts -n istio-system --from-file=ecc/ca-cert.pem \
+    --from-file=ecc/ca-key.pem --from-file=ecc/root-cert.pem \
+    --from-file=ecc/cert-chain.pem \
+    --dry-run -o yaml |
+    kubectl apply -f -
+fi
 
 BASE_CHART="$RELEASE_PATH/manifests/charts/base"
 BASE_NAME=istio-base
@@ -83,8 +85,13 @@ else
   OPENSHIFT=false
 fi
 
+VALIDATION=""
+if [[ ${IN_PLACE_UPGRADE_1_9:-} == "true" ]]; then
+  VALIDATION="--set=global.configValidation=false"
+fi
+
 helm upgrade $BASE_NAME "$BASE_CHART" \
-  --install \
+  --install $VALIDATION \
   --namespace=istio-system "${VALUES_OPTS[@]}" "$@"
 
 if [[ $OPENSHIFT == "true" ]]; then
@@ -101,7 +108,7 @@ if [[ $OPENSHIFT == "true" ]]; then
 fi
 
 if [[ ${PULLSECRET:-} != "" ]]; then
-  kubectl apply -f "$PULLSECRET" -n istio-system
+  kubectl apply -f "$PULLSECRET" --namespace istio-system
 fi
 
 while true; do
@@ -123,11 +130,24 @@ else
     --install \
     --namespace=istio-system "${VALUES_OPTS[@]}" "$@"
 
+  if [[ ${IN_PLACE_UPGRADE_1_9:-} == "true" ]]; then
+    helm upgrade $BASE_NAME "$BASE_CHART" \
+      --set=global.configValidation=true \
+      --namespace=istio-system "${VALUES_OPTS[@]}" "$@"
+  fi
+
   helm upgrade istio-ingress \
     "$RELEASE_PATH/manifests/charts/gateways/istio-ingress" \
     --install \
     --namespace=istio-system "${VALUES_OPTS[@]}" "$@"
 fi
+
+kubectl get namespace --selector=istio-injection=enabled | tail -n +2 | while read -r NS _; do
+  kubectl get deployment --namespace "$NS" -o name | while read -r DEPLOYMENT; do
+    kubectl rollout restart --namespace "$NS" "$DEPLOYMENT"
+    kubectl rollout status --namespace "$NS" "$DEPLOYMENT" --watch=true
+  done
+done
 
 while true; do
   LOAD_BALANCER=$(kubectl get service istio-ingressgateway \
@@ -137,6 +157,8 @@ while true; do
     break
   fi
 done
+
+echo "$LOAD_BALANCER" > load-balancer.value
 
 ANALYSIS_CHART="$RELEASE_PATH/samples/aspenmesh/analysis-emulator"
 if [[ $MINOR_VER == "1.5" ]]; then
@@ -149,40 +171,40 @@ if [ -d "$ANALYSIS_CHART" ]; then
     --namespace=analysis-emulator "${VALUES_OPTS[@]}" "$@"
 fi
 
-kubectl apply -f "$DIR/ready.yaml"
-kubectl apply -f ../private-resources/aspenmesh-istio-private-pr-pull-secret.yaml \
-  --namespace istio-ready
-if [[ $OPENSHIFT == "true" ]]; then
-  kubectl apply -f "$DIR/net-attach-def.yaml" \
+if [[ ${CHECK_READY:-} != "false" ]]; then
+  kubectl apply -f "$DIR/ready.yaml"
+  kubectl apply -f ../private-resources/aspenmesh-istio-private-pr-pull-secret.yaml \
     --namespace istio-ready
-fi
-
-dots() {
-  while [ -f "$TMP_FILE" ]; do
-    echo -n .
-    sleep 10
-  done
-}
-
-echo "http://$LOAD_BALANCER/status/200"
-set +x
-date
-echo "Time: 1 min 2 min 3 min 4 min 5 min 6 min 7 min"
-dots &
-while true; do
-  STATUS=$(curl --silent \
-    --output /dev/null \
-    --write-out "%{http_code}\n" \
-    "http://$LOAD_BALANCER/status/200" || true)
-  if [[ $STATUS == 200 ]]; then
-    break
+  if [[ $OPENSHIFT == "true" ]]; then
+    kubectl apply -f "$DIR/net-attach-def.yaml" \
+      --namespace istio-ready
   fi
-  sleep 5
-done
-echo
-date
-set -x
 
-kubectl delete namespace istio-ready
+  dots() {
+    while [ -f "$TMP_FILE" ]; do
+      echo -n .
+      sleep 10
+    done
+  }
 
-echo "export LOAD_BALANCER=$LOAD_BALANCER"
+  echo "http://$LOAD_BALANCER/status/200"
+  set +x
+  date
+  echo "Time: 1 min 2 min 3 min 4 min 5 min 6 min 7 min"
+  dots &
+  while true; do
+    STATUS=$(curl --silent \
+      --output /dev/null \
+      --write-out "%{http_code}\n" \
+      "http://$LOAD_BALANCER/status/200" || true)
+    if [[ $STATUS == 200 ]]; then
+      break
+    fi
+    sleep 5
+  done
+  echo
+  date
+  set -x
+
+  kubectl delete namespace istio-ready
+fi
